Django_Backend/requirements.txt
django==5.0
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5
django-cors-headers==4.3.1
django-ckeditor==6.7.0
pillow==10.2.0